% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnxietySleep-package.R
\docType{package}
\name{AnxietySleep-package}
\alias{AnxietySleep}
\alias{AnxietySleep-package}
\title{Anxiety & Confinement}
\description{
Data from the anxiety and confinement study from Alvarado et al. (2022) <doi: 10.3390/bs12100398>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NIM-ACh/AnxietySleep}
  \item Report bugs at \url{https://github.com/NIM-ACh/AnxietySleep/issues}
}

}
\author{
\strong{Maintainer}: Matías Castillo-Aguilar \email{matcasti@umag.cl} (\href{https://orcid.org/0000-0001-7291-247X}{ORCID})

Authors:
\itemize{
  \item Caren Alvarado-Aravena \email{caren.ccaa@gmail.com} (\href{https://orcid.org/0000-0002-1851-2499}{ORCID})
  \item Cristian Núñez-Espinosa \email{cristian.nunez@umag.cl} (\href{https://orcid.org/0000-0002-9896-7062}{ORCID})
}

}
\keyword{internal}
