% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{Datos del estudio de ansiedad, calidad de sueño y confinamiento}
\format{
Un data frame con 625 filas y 6 variables:

\itemize{
\item{\code{edad}: Número entero. Edad de los participantes en años.}
\item{\code{sexo}: Factor c/ 2 niveles "Mujer", "Hombre". Sexo de los participantes.}
\item{\code{zona}: Factor c/ 2 niveles "EC", "SC". Zona en la que se encontraba de residente
el participante al momento de ser encuestado.}
\item{\code{beck_global}: Número entero. Puntaje global del Inventario de Ansiedad de Beck.}
\item{\code{pits_global}: Número entero. Puntaje global cuestionario de calidad de sueño de
Pittsburgh.}
\item{\code{cat_edad}: Factor c/ 4 niveles "18-25", "26-40", "41-50", ">50" . Edad de los
participantes en años.}
}
}
\source{
Alvarado et al, 2021.
}
\usage{
dataset
}
\description{
Herramientas complementarias para el análisis de los datos del estudio de
ansiedad, calidad de sueño y confinamiento (2021).
}
\keyword{datasets}
